feat(types): cria pacote para compartilhar types
