+++ apps/admin/package.json
@@ -15,7 +15,7 @@
     "@mantine/colors-generator": "^7.8.0",
     "@mantine/core": "^7.8.0",
     "@mantine/hooks": "^7.8.0",
-    "@repo/components": "workspace:*",
+    "@repo/components": "workspace:^",
     "@repo/constants": "workspace:^",
     "@tabler/icons-react": "^3.2.0",
     "axios": "^1.6.8",
@@ -30,8 +30,9 @@
   },
   "devDependencies": {
     "@next/eslint-plugin-next": "^14.1.1",
-    "@repo/eslint-config": "workspace:*",
-    "@repo/typescript-config": "workspace:*",
+    "@repo/eslint-config": "workspace:^",
+    "@repo/types": "workspace:^",
+    "@repo/typescript-config": "workspace:^",
     "@swc/core": "^1.4.13",
     "@swc/jest": "^0.2.36",
     "@testing-library/jest-dom": "^6.4.2",
