+++ apps/web/package.json
@@ -17,7 +17,7 @@
     "@mantine/core": "^7.8.0",
     "@mantine/hooks": "^7.8.0",
     "@mantine/notifications": "^7.8.0",
-    "@repo/components": "workspace:*",
+    "@repo/components": "workspace:^",
     "@repo/constants": "workspace:^",
     "@supabase/supabase-js": "^2.42.3",
     "@tabler/icons-react": "^3.2.0",
@@ -35,8 +35,9 @@
   },
   "devDependencies": {
     "@next/eslint-plugin-next": "^14.1.1",
-    "@repo/eslint-config": "workspace:*",
-    "@repo/typescript-config": "workspace:*",
+    "@repo/eslint-config": "workspace:^",
+    "@repo/types": "workspace:^",
+    "@repo/typescript-config": "workspace:^",
     "@swc/core": "^1.4.13",
     "@swc/jest": "^0.2.36",
     "@testing-library/jest-dom": "^6.4.2",
