+++ packages/types/src/account.ts
@@ -1,0 +1,11 @@
+export type Account = {
+  id: number;
+  balance: number;
+  benefits: {
+    id: number;
+    name: string;
+    color_from: string;
+    color_to: string;
+    icon: string;
+  };
+};
