+++ packages/types/src/transaction.ts
@@ -1,0 +1,11 @@
+export type Transaction = {
+  id: number;
+  amount: number;
+  name: string;
+  created_at: string;
+  accounts: {
+    benefits: {
+      icon: string;
+    };
+  };
+};
